Moves k8s environment check to new common file.
